combine python tools into one single odrvtool.py
