+++ tools/odrive/dfuse/__init__.py
@@ -1,0 +1,4 @@
+from .DfuDevice import DfuDevice
+from .DfuStatus import DfuStatus
+from .DfuState import DfuState
+from .DfuFile import DfuFile
