+++ tools/odrive/enums.py
@@ -1,0 +1,11 @@
+
+# some enums described in the README
+# TODO: This is dangerous. Transmit as part of the JSON
+MOTOR_TYPE_HIGH_CURRENT = 0
+#MOTOR_TYPE_LOW_CURRENT = 1
+MOTOR_TYPE_GIMBAL = 2
+
+CTRL_MODE_VOLTAGE_CONTROL = 0,
+CTRL_MODE_CURRENT_CONTROL = 1,
+CTRL_MODE_VELOCITY_CONTROL = 2,
+CTRL_MODE_POSITION_CONTROL = 3
